requirements.txt
networkx>=3,<4
numpy>=1
pygame>=2
sympy>=1
colorama>=0.4.6
matplotlib>=3
tqdm>=4
PyOpenGL>=3.1.7
scipy>=1.7
Pillow>=10
fonttools>=4
scikit-image>=0.21
